+++ tests/testthat/test-beastier_report.R
@@ -1,0 +1,3 @@
+test_that("use", {
+  expect_output(beastier_report())
+})