+++ NEWS.md
@@ -9,6 +9,7 @@ Newest versions at top.
   * Add `rename_beast2_options_filenames`
   * Add `get_beast2_options_filenames`
   * Allow tildes in file paths, for example `~/beast2.xml`
+  * Add `beastier_report` to help create a bug report
 
 ### MINOR IMPROVEMENTS